put timemap data processor to dedicated file
