+++ src/js/index.js
@@ -1,4 +1,8 @@
+import * as d3 from 'd3';
+import {processTimeMapData} from './processing/timemap';
+
 /**
+ *
  * Radial Tree Library
  *
  * @param {DOMElement} element
@@ -8,10 +12,8 @@
  &collapse=timestamp:4&limit=100000``.
  * @param {Object} option
  * Option baseURL defines the target Wayback Machine server.
+ *
  */
-import * as d3 from 'd3';
-
-
 export function RadialTree(element, cdx_data, option) {
   let GlobYear = 0;
   let baseURL = 'https://web.archive.org';
@@ -23,7 +25,7 @@ export function RadialTree(element, cdx_data, option) {
 
   init(element);
 
-  let {allYears, yearData} = getData(option.url, cdx_data);
+  let {allYears, yearData} = processTimeMapData(option.url, cdx_data);
 
   createYearButtons(element, option, allYears, yearData);
 
@@ -43,72 +45,6 @@ export function RadialTree(element, cdx_data, option) {
     content.style.display = 'block';
 
     container.appendChild(content);
-  }
-
-  function getData(url, response, cb) {
-    let regexHTTP = /http:\/\//;
-    let regexHTTPS = /https:\/\//;
-    let regexLast = /\/$/;
-    url.replace(regexHTTP, '');
-    url.replace(regexHTTPS, '');
-    url.replace(regexLast, '');
-
-    if (response.length === 0) {
-      return {
-        allYears: [],
-        yearData: [],
-      };
-    }
-
-    let yearUrl = [];
-    for (let i = 1; i < response.length; i++) {
-      if (response[i][1].match(/jpg|pdf|png|form|gif/)) {
-        continue;
-      }
-      response[i][1] = response[i][1].trim().replace(':80/', '/');
-      if (response[i][0] in yearUrl) {
-        yearUrl[response[i][0]].push(response[i][1]);
-      } else {
-        yearUrl[response[i][0]] = [response[i][1]];
-      }
-    }
-    let ret = [];
-    for (let year in yearUrl) {
-      ret.push([year].concat(yearUrl[year]));
-    }
-    /** ret has the following format:
-     *  array(
-     *    array(2005, url1, url2, .... urlN),
-     *    ...
-     *  ) **/
-    let years = (function () {
-      for (let i = 0; i < ret.length; i++) {
-        for (let j = 1; j < ret[i].length; j++) {
-          let url;
-          if (ret[i][j].includes('http')) {
-            url = ret[i][j].substring(7);
-          } else if (ret[i][j].includes('https')) {
-            url = ret[i][j].substring(8);
-          }
-          if (url.includes('//')) {
-            url = url.split('//').join('/');
-          }
-          url = url.split('/').join('/');
-          ret[i][j] = url;
-        }
-      }
-      return ret;
-    }());
-    let all_years = years.map(function (year) {
-      if (year.length > 1) {
-        return year[0];
-      }
-    });
-
-    return {
-      allYears: all_years,
-      yearData: years,
-    };
   }
 
   function createYearButtons(element, option, allYears, yearData) {
